Traktor: Implemented a "hash vector" in Model, used for vertices to improve "add unique" performance. Rearranged all cache directories, using a more homogenous structure. Optimized shader operations, no longer create a deep clone for several operations. Added "Flipboard" widget from GearUp into UiKit.
[git-p4: depot-paths = "//depot/traktor/": change = 24486]

+++ code/Model/Operations/Reduce.cpp
@@ -122,7 +122,7 @@ Vector4 triangleTipPoint(const Model& model, const ModelAdjacency& adjacency, ui
 	}
 
 	if (count > 0)
-		tipPoint = (tipPoint / Scalar(count)).xyz1();
+		tipPoint = (tipPoint / Scalar((float)count)).xyz1();
 	else
 		tipPoint = midPoint;
